# Implement HLE macro for DrawElementsIndirect