--- a/Ryujinx.Graphics.Gpu/Engine/MME/MacroHLEFunctionName.cs
+++ b/Ryujinx.Graphics.Gpu/Engine/MME/MacroHLEFunctionName.cs
@@ -8,6 +8,7 @@
         None,
         ClearColor,
         ClearDepthStencil,
+        DrawElementsIndirect,
         MultiDrawElementsIndirectCount
     }
 }