--- a/Ryujinx.Graphics.Gpu/Engine/MME/MacroHLEFunctionName.cs
+++ b/Ryujinx.Graphics.Gpu/Engine/MME/MacroHLEFunctionName.cs
@@ -8,6 +8,8 @@
         None,
         ClearColor,
         ClearDepthStencil,
+        DrawArraysInstanced,
+        DrawElementsInstanced,
         DrawElementsIndirect,
         MultiDrawElementsIndirectCount
     }