… Ryujinx.Ava/Assets/Locales/en_US.json
… Ryujinx.Ava/Assets/Locales/pt_BR.json
… Ryujinx.Graphics.GAL/Multithreading/CommandType.cs
… Ryujinx.Graphics.Gpu/Engine/MME/MacroHLETable.cs
… Ryujinx.Graphics.Gpu/Engine/Threed/DrawManager.cs
… Ryujinx.Graphics.Gpu/Engine/Threed/IndirectDrawType.cs
… Ryujinx.Graphics.Gpu/Engine/Threed/StateUpdater.cs
… Ryujinx.Graphics.Gpu/Engine/Threed/ThreedClass.cs
… Ryujinx.Graphics.Gpu/Memory/BufferCache.cs
… Ryujinx.Graphics.Gpu/Shader/DiskCache/DiskCacheGpuAccessor.cs
… Ryujinx.Graphics.Gpu/Shader/DiskCache/DiskCacheHostStorage.cs
… Ryujinx.Graphics.Gpu/Shader/GpuAccessor.cs
… Ryujinx.Graphics.Gpu/Shader/GpuChannelGraphicsState.cs
… Ryujinx.Graphics.Gpu/Shader/ShaderCache.cs
… Ryujinx.Graphics.Gpu/Shader/ShaderSpecializationList.cs
… Ryujinx.Graphics.Gpu/Shader/ShaderSpecializationState.cs
… Ryujinx.Graphics.OpenGL/Pipeline.cs
… Ryujinx.Graphics.Shader/CodeGen/Glsl/Declarations.cs
… Ryujinx.Graphics.Shader/CodeGen/Glsl/OperandManager.cs
… Ryujinx.Graphics.Shader/CodeGen/Spirv/Declarations.cs
… Ryujinx.Graphics.Shader/CodeGen/Spirv/SpirvGenerator.cs
… Ryujinx.Graphics.Shader/Constants.cs
… Ryujinx.Graphics.Shader/IGpuAccessor.cs
… Ryujinx.Graphics.Shader/ShaderProgramInfo.cs
… Ryujinx.Graphics.Shader/Translation/AttributeConsts.cs
… Ryujinx.Graphics.Shader/Translation/AttributeInfo.cs
… Ryujinx.Graphics.Shader/Translation/FeatureFlags.cs
… Ryujinx.Graphics.Shader/Translation/Rewriter.cs
… Ryujinx.Graphics.Shader/Translation/Translator.cs
… Ryujinx.Graphics.Vulkan/BufferHolder.cs