This implements an HLE macro for the `DrawElementsIndirect` NVN macro, improving performance on games that uses this type of draw. This is mostly extracted from #3306.

**Affected games:**

Nier Automata: The End of YoRHa Edition. Lower FIFO % and slightly better performance in some areas.
Before:
![image](https://user-images.githubusercontent.com/5624669/194710931-cbfbd93a-8578-4213-b86c-4de6c6a6f4da.png)
After:
![image](https://user-images.githubusercontent.com/5624669/194710948-51cbc6ce-7508-4cc4-8d0c-720190084e5e.png)

Monster Hunter Rise. Lower FIFO % and better performance in some areas (hard to tell how much better without unlocking the frame rate).
Before:
![image](https://user-images.githubusercontent.com/5624669/194711610-8c0151b0-1c13-43c8-8ebb-bec90b5d9924.png)
After:
![image](https://user-images.githubusercontent.com/5624669/194711060-39d2223a-0a28-4070-a271-8dcc63ebed04.png)

Marvel Ultimate Alliance 3. Unfortunately this one won't really be improved since it also uses conditional rendering with GPU written buffers, so it still needs to flush data for that which nullifies all benefits from this change. I'm including it here just to indicate that it needs to be tested to make sure its not *worse*.

Nintendo Switch Sports, but this one is not playable right now due to other issues.

**Problems:**

- Conditional rendering comparison currently does not flush buffers before reading values. Marvel Ultimate Alliance 3 needs this, but since the counter values are in the same buffer as the indirect data, it was being flushed with the indirect data before. Since this change makes it skip this flush, I had to make the counter reads trigger flush to avoid regressing this game.
- We have no way to know the index buffer size with indirect draws, since the sub-range that is used is part of the indirect data (`firstIndex` and `indexCount`). This creates various problems and complications.
  - We don't know how much buffer data we need to update for the draw. Right now it just assumes a maximum size and updates all that.
  - ~~Can't convert index buffers on Vulkan, so any games using indirect draws with quads won't work properly. This can be fixed by doing the conversion in compute while reading the range from the indirect data there too.~~ This is now supported on 711ce42d176948a6e93127033ea6bccc09b7cd31.

**Other changes:**

- `MultiDrawIndirectCount` and `DrawIndirectCount` methods were combined to avoid code duplication.
- None of the indirect draw methods reads the CPU written indirect data anymore, and instead just always assume a fixed index buffer size. That's because if the count is written from GPU, then the CPU side data might contain garbage.
- The `MultiDraw*` methods were renamed removing the `Multi` prefix. This matches the Vulkan API names and allows all draw methods to be grouped together when sorted alphabetically, which is pretty neat.

Testing is welcome.