@GamerzHell9137 2022-10-08
Marvel seems to run fine, can't really see any difference but maybe a low end system might see something.