@gdkchan 2022-10-15
We got a report on Discord of a regression on Mario Golf Super Rush with the previous change. It was because the game uses indirect draw with quads, which is not supported on Vulkan, and we did not support the index buffer conversion in this case because it requires having access to the index count and first index on CPU, which for indirect draw we don't as its stored on the indirect buffer. So to support this case, I implemented index buffer conversion on compute. It works like this:
- First dispatch reads indirect data and determines the index buffer bounds, it then writes the new indirect data with the modified index count and first index. The primitive count is written into a buffer to be used as parameter for the second dispatch, which is indirect.
- Second dispatch is indirect, X is written in the step above, Y and Z are always 1. It has one invocation per primitive, so each invocation will copy the indices for that primitive following the pattern specified for the topology conversion.

This also introduces "dependencies" on the `CacheByRange`. The index buffer being valid "depends" on the indirect buffer not being modified. So if the indirect buffer is modified, the cached index buffer has to be removed too. And for multi-draw, the indirect data being valid depends on the draw count not being modified, since it is used to determine the bounds of the indirect buffer. Now each entry on `CacheByRange` also stores a dependency list. When the entry is disposed, it also forces the removal of all the dependencies from their caches, which should cause them to be disposed too and cascade.

Testing is welcome.