@gdkchan 2022-10-20
The assumed buffer range is not large enough for some games, so I will just wait for #3775 and try updating the whole buffer then.