@gdkchan 2022-11-01
This is ready for review again. For the index buffer size problem, it not uses the contiguous mapped size starting at the index buffer address as the size, but it still has a maximum size that it restricts to. That is to avoid the buffer/region being checked being too large.

Needs to be tested on games that uses indirect draws to ensure that they did not regress (rendering is still correct etc):
- Nier Automata.
- Mario Golf: Super Rush.
- Monster Hunter Rise.
- Ghosts 'n Goblins: Resurrection.
- Bayonetta 3.
- Marvel Ultimate Alliance 3.
In addition to that, we need to make sure performance on those games did not regress compared to master:
- Monster Hunter Rise.
- Bayonetta 3.