@S0gnat0re 2022-11-02
Found a consistent crash on bayonetta 3 that isn't present on master i believe

> 00:01:01.086 |S| HLE.OsThread.16 ServiceAm SetIdleTimeDetectionExtension: Stubbed. {idleTimeDetectionExtension: 0}
00:01:02.349 |E| HLE.OsThread.8 Application : Unhandled exception caught: System.Reflection.TargetInvocationException: Exception has been thrown by the target of an invocation.
 ---> LibHac.Common.HorizonResultException: ResultFsNonRealDataVerificationFailed (2002-4604): Hash error!
   at LibHac.Common.ThrowHelper.ThrowResult(Result result, String message)
   at LibHac.Tools.FsSystem.IntegrityVerificationStorage.ReadImpl(Int64 offset, Span`1 destination, IntegrityCheckLevel integrityCheckLevel)
   at LibHac.Tools.FsSystem.CachedStorage.ReadBlock(CacheBlock block, Int64 index)
   at LibHac.Tools.FsSystem.CachedStorage.GetBlock(Int64 blockIndex)
   at LibHac.Tools.FsSystem.CachedStorage.Read(Int64 offset, Span`1 destination)
   at LibHac.Tools.FsSystem.StorageStream.Read(Byte[] buffer, Int32 offset, Int32 count)
   at System.IO.Stream.Read(Span`1 )
   at LibHac.Tools.FsSystem.StreamStorage.Read(Int64 offset, Span`1 destination)
   at LibHac.FsSrv.Impl.StorageInterfaceAdapter.Read(Int64 offset, OutBuffer destination, Int64 size)
   at Ryujinx.HLE.HOS.Services.Fs.FileSystemProxy.IStorage.Read(ServiceCtx context)
   --- End of inner exception stack trace ---
   at System.RuntimeMethodHandle.InvokeMethod(Object , Span`1& , Signature , Boolean , Boolean )
   at System.Reflection.RuntimeMethodInfo.Invoke(Object , BindingFlags , Binder , Object[] , CultureInfo )
   at Ryujinx.HLE.HOS.Services.IpcService.CallHipcMethod(ServiceCtx context)
   at Ryujinx.HLE.HOS.Services.ServerBase.Process(Int32 serverSessionHandle, UInt64 recvListAddr)
   at Ryujinx.HLE.HOS.Services.ServerBase.ServerLoop()
   at Ryujinx.HLE.HOS.Services.ServerBase.Main()
   at Ryujinx.HLE.HOS.Kernel.Threading.KThread.ThreadStart()
   at System.Threading.Thread.StartCallback()

[Ryujinx_1.1.0+03843f9_2022-11-02_01-40-54.log](https://github.com/Ryujinx/Ryujinx/files/9914946/Ryujinx_1.1.0%2B03843f9_2022-11-02_01-40-54.log)

Savefile to replicate
(keep running forward, after a cutscene an enemy will start aiming and the crash should happen almost instantly)
[0.zip](https://github.com/Ryujinx/Ryujinx/files/9914959/0.zip)

Aside this the game renders normally and seems to stutter less as far as i can tell