@Bjorn29512 2022-11-02
Bayonetta 3 performance seems fine so far : 

Master : 
![image](https://user-images.githubusercontent.com/110204265/199470356-51833537-8c8d-4fbb-a42a-269068c512dd.png)
Average framerate  :   59.1 FPS
                     Minimum framerate  :   56.2 FPS
                     Maximum framerate  :   62.7 FPS
                     1% low framerate   :   43.0 FPS
                     0.1% low framerate :   38.1 FPS
                     
PR :
![image](https://user-images.githubusercontent.com/110204265/199472129-1ce549d8-9389-48ab-83bc-755992d5f2bd.png)
 Average framerate  :   59.2 FPS
                     Minimum framerate  :   56.1 FPS
                     Maximum framerate  :   63.4 FPS
                     1% low framerate   :   43.6 FPS
                     0.1% low framerate :   35.7 FPS